User Defaults Set
Made a class PetRepository
gave it a key - PET_KEY
I initialized the data to check if you have a pet already. If not it results to a dog. We will be changing that to be defaulting to spin the gum ball machine.
made a func called load data that we will call when the game gets initialized/opened
when a change is made in the game like time spent in the game or we press a button we can call saveData so everything that the user has done gets saved to the UserDefaults

Also added a calculateTimeSince function that will keep track of when we last fed, showered, played etc... that returns seconds since that time